requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.26.0
python-dotenv==0.19.0